--- yyGang/src/main/java/com/beyond3/yyGang/controller/UserController.java
@@ -1,4 +1,0 @@
-package com.beyond3.yyGang.controller;
-
-public class UserController {
-}
