--- yyGang/src/main/java/com/beyond3/yyGang/domain/member/Gender.java
@@ -1,5 +1,0 @@
-package com.beyond3.yyGang.domain.member;
-
-public enum Gender {
-    Male, Female
-}
