--- yyGang/src/main/java/com/beyond3/yyGang/domain/member/Role_name.java
@@ -1,5 +1,0 @@
-package com.beyond3.yyGang.domain.member;
-
-public enum Role_name {
-    CUSTOMER, SELLER, PHARMACIST, ADMIN
-}
