--- yyGang/src/main/java/com/beyond3/yyGang/domain/store/PayStatus.java
@@ -1,5 +1,0 @@
-package com.beyond3.yyGang.domain.store;
-
-public enum PayStatus {
-    WAITING, FAIL, SUCCESS
-}
