+++ yyGang/src/main/java/com/beyond3/yyGang/member/domain/Gender.java
@@ -1,0 +1,5 @@
+package com.beyond3.yyGang.member.domain;
+
+public enum Gender {
+    Male, Female
+}
