+++ yyGang/src/main/java/com/beyond3/yyGang/member/domain/Role_name.java
@@ -1,0 +1,5 @@
+package com.beyond3.yyGang.member.domain;
+
+public enum Role_name {
+    CUSTOMER, SELLER, PHARMACIST, ADMIN
+}
