+++ yyGang/src/main/java/com/beyond3/yyGang/store/domain/Cart.java
@@ -1,0 +1,33 @@
+package com.beyond3.yyGang.store.domain;
+
+import com.beyond3.yyGang.member.domain.Users;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+import lombok.Getter;
+
+import java.util.List;
+
+@Entity
+@Getter
+@Table(name = "cart")
+public class Cart {
+
+    @Id
+    @GeneratedValue
+    @Column(name = "cart_id")
+    private Long cartId;
+
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private Users user;
+
+    @OneToMany(mappedBy = "cart")
+    private List<CartOption> cartOptions;
+}
